[VE] Change to not compile libomptarget for VE
Change to not compile libomptarget for VE at the moment since it is
not compilable after https://reviews.llvm.org/D129875.  I also made
libomptarget to compile openmp and libomptarget for VE for the case
of improvements this.  Type `make libomptarget` to compile them for
VE although this cause compile errors..